feat: :art: WIP

+++ users/src/services/user-service.ts
@@ -1,22 +1,36 @@
 import { UserModel } from '../types';
 import User from '../models/user';
 import { Op } from 'sequelize';
+import { EventUsers } from './queue/EventUsers';
+import { EventType } from '../events/EventUserDomain';
 
 export default class UserService {
-  public createUser = async (userData: User) => {
-    console.log(userData);
 
-    if (
-      await User.findOne({
-        where: {
-          [Op.or]: [{ username: userData.username }, { email: userData.email }],
-        },
-      })
-    ) {
+  private eventUsers: EventUsers;
+
+  constructor() {
+    this.eventUsers = new EventUsers();
+  }
+
+  public createUser = async (userData: User) => {
+    const existingUser = await User.findOne({
+      where: {
+        [Op.or]: [{ username: userData.username }, { email: userData.email }],
+      },
+    });
+
+    if (existingUser) {
       throw new Error('Usuário já cadastrado');
     } else {
       const user = await User.create(userData);
-      user.save();
+
+      console.log("DISPARANDO EVENTO DE CRIAÇÃO");
+      this.eventUsers.publisher({
+        username: user.dataValues.username,
+        email: user.dataValues.email,
+        referenceId: user.dataValues.referenceId,
+        eventType: EventType.CREATED
+      });
       return user;
     }
   };
@@ -27,35 +41,14 @@ export default class UserService {
 
   public updateUser = async (userNewData: UserModel) => {
     try {
-      const userFromDb = await User.findOne({
-        where: {
-          id: userNewData.id,
-        },
-      });
-      if (!userFromDb) throw new Error('Usuário não encontrado');
-      const user = {
-        email: userNewData.email ? userNewData.email : userFromDb.email,
-        username: userNewData.username
-          ? userNewData.username
-          : userFromDb.username,
-        password: userNewData.password
-          ? userNewData.password
-          : userFromDb.password,
-        firstname: userNewData.firstname
-          ? userNewData.firstname
-          : userFromDb.firstName,
-        lastname: userNewData.lastname
-          ? userNewData.lastname
-          : userFromDb.lastName,
-      };
 
-      return await User.update(
+      const [rowsUpdated] = await User.update(
         {
-          email: user.email,
-          username: user.username,
-          firstName: user.firstname,
-          lastName: user.lastname,
-          password: user.password,
+          email: userNewData.email,
+          username: userNewData.username,
+          firstName: userNewData.firstname,
+          lastName: userNewData.lastname,
+          password: userNewData.password,
         },
         {
           where: {
@@ -63,16 +56,55 @@ export default class UserService {
           },
         }
       );
+
+      const user = await User.findByPk(userNewData.id);
+
+      if (user == null) {
+        console.log("usuario nao encontrado com o id:", userNewData.id)
+      }
+
+      if (rowsUpdated === 0) {
+        console.log("Usuario nao atualizou os campos")
+      }
+
+      if (user && rowsUpdated > 0) {
+        console.log("DISPARANDO EVENTO DE ATUALIZAÇÃO");
+        this.eventUsers.publisher({
+          username: userNewData.username,
+          email: userNewData.email,
+          referenceId: user.dataValues.referenceId,
+          eventType: EventType.UPDATED
+        });
+      }
+
+      return userNewData;
     } catch (error) {
       console.error('erro no servidor => ', error);
     }
   };
 
-  public deleteUser = async (userId: number) => {
-    return await User.destroy({
+  public deleteUser = async (userId: number): Promise<boolean> => {
+    const user = await User.findByPk(userId);
+
+    if (!user) {
+      throw new Error('Usuário não encontrado');
+    }
+
+    await User.destroy({
       where: {
         id: userId,
       },
     });
+
+    console.log("Usuário removido");
+
+    this.eventUsers.publisher({
+      username: user.dataValues.username,
+      email: user.dataValues.email,
+      referenceId: user.dataValues.referenceId,
+      eventType: EventType.DELETED
+    });
+
+    return true
   };
 }
